migrate api_call, fix tests
